# Throws an exception if the width/height of the entity is invalid