--- a/src/pocketmine/entity/Entity.php
+++ b/src/pocketmine/entity/Entity.php
@@ -610,6 +610,10 @@ abstract class Entity extends Location implements Metadatable, EntityIds{
 
 		$this->fallDistance = $this->namedtag->getFloat("FallDistance", 0.0);
 
+		if(!isset($this->width) or !isset($this->height)){
+			throw new \InvalidStateException(static::class . " has invalid width/height");
+		}
+
 		$this->propertyManager = new DataPropertyManager();
 
 		$this->propertyManager->setLong(self::DATA_FLAGS, 0);