--- a/src/pocketmine/entity/Entity.php
+++ b/src/pocketmine/entity/Entity.php
@@ -611,7 +611,7 @@ abstract class Entity extends Location implements Metadatable, EntityIds{
 		$this->fallDistance = $this->namedtag->getFloat("FallDistance", 0.0);
 
 		if(!isset($this->width) or !isset($this->height)){
-			throw new \RuntimeException("Entity " . static::class . " must declare the properties height and width and must be a float.");
+			throw new \RuntimeException("Entity " . static::class . " must declare height and width properties and must be a float.");
 		}
 
 		$this->propertyManager = new DataPropertyManager();