@DaisukeDaisuke 2021-05-07
I'm sorry.
I accidentally clicked the rereview button.
There is no problem with ignoring the specification.