@dktapps 2021-06-09
I'm doubtful if it's worth merging this considering PM4 already has a better solution for this problem (abstract getInitialSizeInfo()).